backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
fastapi==0.115.5
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
sqlalchemy==2.0.27
psycopg2-binary
python-dotenv==1.0.0
